requirements.txt
# Grace Irvine Ministry Scheduler Dependencies
# 支持本地和云环境部署

# 数据处理 / Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Google Cloud 集成 / Google Cloud Integration
gspread==5.12.0
google-auth==2.23.4
google-cloud-secret-manager==2.16.4

# Web 界面 / Web Interface
streamlit==1.28.1

# 配置管理 / Configuration Management
python-dotenv==1.0.0

# 日期处理 / Date Processing
python-dateutil==2.8.2

# 邮件发送 / Email Sending
jinja2==3.1.2

# 配置文件解析 / Configuration Parsing
pyyaml==6.0.1

# HTTP 客户端 / HTTP Client (for health checks)
requests==2.31.0

# ICS日历支持 / ICS Calendar Support
icalendar==5.0.11
pytz==2023.3
schedule==1.2.0

# 后台API服务 / Background API Service
flask==3.0.0
flask-cors==4.0.0

# Google Cloud Storage / Cloud Storage
google-cloud-storage==2.10.0